Unify pre/post-boot unsupported operation errors
Also move back Vmm specific errors to the Vmm file.
