--- src/vmm/src/error.rs
@@ -1,549 +1,0 @@
-// Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-//! Contains error related logic.
-
-use std::fmt::{Display, Formatter};
-use std::io;
-
-use device_manager;
-use logger::error::LoggerError;
-use polly::event_manager;
-use vstate;
-
-/// Errors associated with the VMM internal logic. These errors cannot be generated by direct user
-/// input, but can result from bad configuration of the host (for example if Firecracker doesn't
-/// have permissions to open the KVM fd).
-#[derive(Debug)]
-pub enum Error {
-    /// This error is thrown by the minimal boot loader implementation.
-    ConfigureSystem(arch::Error),
-    /// Legacy devices work with Event file descriptors and the creation can fail because
-    /// of resource exhaustion.
-    #[cfg(target_arch = "x86_64")]
-    CreateLegacyDevice(device_manager::legacy::Error),
-    /// An operation on the epoll instance failed due to resource exhaustion or bad configuration.
-    EpollFd(io::Error),
-    /// Cannot read from an Event file descriptor.
-    EventFd(io::Error),
-    /// An event arrived for a device, but the dispatcher can't find the event (epoll) handler.
-    DeviceEventHandlerNotFound,
-    /// An epoll handler can't be downcasted to the desired type.
-    DeviceEventHandlerInvalidDowncast,
-    /// Cannot access kernel file.
-    KernelFile(io::Error),
-    /// Cannot open /dev/kvm. Either the host does not have KVM or Firecracker does not have
-    /// permission to open the file descriptor.
-    KvmContext(vstate::Error),
-    #[cfg(target_arch = "x86_64")]
-    /// Cannot add devices to the Legacy I/O Bus.
-    LegacyIOBus(device_manager::legacy::Error),
-    /// Cannot load command line.
-    LoadCommandline(kernel::cmdline::Error),
-    /// Internal logger error.
-    Logger(LoggerError),
-    /// I8042 Error.
-    I8042Error(devices::legacy::I8042DeviceError),
-    /// Epoll wait failed.
-    Poll(io::Error),
-    /// Cannot add a device to the MMIO Bus.
-    RegisterMMIODevice(device_manager::mmio::Error),
-    /// Cannot build seccomp filters.
-    SeccompFilters(seccomp::Error),
-    /// Write to the serial console failed.
-    Serial(io::Error),
-    /// Cannot set mode for terminal.
-    StdinHandle(utils::errno::Error),
-    /// Cannot create Timer file descriptor.
-    TimerFd(io::Error),
-    /// Vcpu error.
-    Vcpu(vstate::Error),
-    /// Cannot spawn a new Vcpu thread.
-    VcpuSpawn(std::io::Error),
-    /// Vm error.
-    Vm(vstate::Error),
-    /// Polly error wrapper.
-    EventManager(event_manager::Error),
-}
-
-impl Display for Error {
-    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
-        use self::Error::*;
-
-        match self {
-            ConfigureSystem(e) => write!(f, "System configuration error: {:?}", e),
-            #[cfg(target_arch = "x86_64")]
-            CreateLegacyDevice(e) => write!(f, "Error creating legacy device: {:?}", e),
-            EpollFd(e) => write!(f, "Epoll fd error: {}", e),
-            EventFd(e) => write!(f, "Event fd error: {}", e),
-            DeviceEventHandlerNotFound => write!(
-                f,
-                "Device event handler not found. This might point to a guest device driver issue."
-            ),
-            DeviceEventHandlerInvalidDowncast => write!(
-                f,
-                "Device event handler couldn't be downcasted to expected type."
-            ),
-            KernelFile(e) => write!(f, "Cannot access kernel file: {}", e),
-            KvmContext(e) => write!(f, "Failed to validate KVM support: {:?}", e),
-            #[cfg(target_arch = "x86_64")]
-            LegacyIOBus(e) => write!(f, "Cannot add devices to the legacy I/O Bus. {}", e),
-            LoadCommandline(e) => write!(f, "Cannot load command line: {}", e),
-            Logger(e) => write!(f, "Logger error: {}", e),
-            I8042Error(e) => write!(f, "I8042 error: {}", e),
-            Poll(e) => write!(f, "Epoll wait failed: {}", e),
-            RegisterMMIODevice(e) => write!(f, "Cannot add a device to the MMIO Bus. {}", e),
-            SeccompFilters(e) => write!(f, "Cannot build seccomp filters: {}", e),
-            Serial(e) => write!(f, "Error writing to the serial console: {:?}", e),
-            StdinHandle(e) => write!(f, "Failed to set mode for terminal: {}", e),
-            TimerFd(e) => write!(f, "Error creating timer fd: {}", e),
-            Vcpu(e) => write!(f, "Vcpu error: {}", e),
-            VcpuSpawn(e) => write!(f, "Cannot spawn Vcpu thread: {}", e),
-            Vm(e) => write!(f, "Vm error: {}", e),
-            EventManager(e) => write!(f, "Event manager error: {:?}", e),
-        }
-    }
-}
-
-/// Shorthand result type for internal VMM commands.
-pub type Result<T> = std::result::Result<T, Error>;
-
-/*
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use kernel;
-
-    // Helper function to get ErrorKind of error.
-    fn error_kind<T: std::convert::Into<VmmActionError>>(err: T) -> ErrorKind {
-        let err: VmmActionError = err.into();
-        err.kind().clone()
-    }
-
-    #[test]
-    fn test_drive_error_conversion() {
-        // Test `DriveError` conversion
-        assert_eq!(
-            error_kind(DriveError::CannotOpenBlockDevice(
-                io::Error::from_raw_os_error(0)
-            )),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(DriveError::InvalidBlockDevicePath),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(DriveError::BlockDevicePathAlreadyExists),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(DriveError::BlockDeviceUpdateFailed),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(DriveError::OperationNotAllowedPreBoot),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(DriveError::UpdateNotAllowedPostBoot),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(DriveError::RootBlockDeviceAlreadyAdded),
-            ErrorKind::User
-        );
-    }
-
-    #[test]
-    fn test_vmconfig_error_conversion() {
-        // Test `VmConfigError` conversion
-        assert_eq!(error_kind(VmConfigError::InvalidVcpuCount), ErrorKind::User);
-        assert_eq!(
-            error_kind(VmConfigError::InvalidMemorySize),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(VmConfigError::UpdateNotAllowedPostBoot),
-            ErrorKind::User
-        );
-    }
-
-    #[test]
-    fn test_network_interface_error_conversion() {
-        // Test `NetworkInterfaceError` conversion
-        assert_eq!(
-            error_kind(NetworkInterfaceError::GuestMacAddressInUse(String::new())),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(NetworkInterfaceError::EpollHandlerNotFound(
-                Error::DeviceEventHandlerNotFound
-            )),
-            ErrorKind::Internal
-        );
-        assert_eq!(
-            error_kind(NetworkInterfaceError::HostDeviceNameInUse(String::new())),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(NetworkInterfaceError::DeviceIdNotFound),
-            ErrorKind::User
-        );
-        // NetworkInterfaceError::OpenTap can be of multiple kinds.
-        {
-            assert_eq!(
-                error_kind(NetworkInterfaceError::OpenTap(
-                    utils::net::TapError::OpenTun(io::Error::from_raw_os_error(0))
-                )),
-                ErrorKind::User
-            );
-            assert_eq!(
-                error_kind(NetworkInterfaceError::OpenTap(
-                    utils::net::TapError::CreateTap(io::Error::from_raw_os_error(0))
-                )),
-                ErrorKind::User
-            );
-            assert_eq!(
-                error_kind(NetworkInterfaceError::OpenTap(
-                    utils::net::TapError::IoctlError(io::Error::from_raw_os_error(0))
-                )),
-                ErrorKind::Internal
-            );
-            assert_eq!(
-                error_kind(NetworkInterfaceError::OpenTap(
-                    utils::net::TapError::CreateSocket(io::Error::from_raw_os_error(0))
-                )),
-                ErrorKind::Internal
-            );
-            assert_eq!(
-                error_kind(NetworkInterfaceError::OpenTap(
-                    utils::net::TapError::InvalidIfname
-                )),
-                ErrorKind::User
-            );
-        }
-        assert_eq!(
-            error_kind(NetworkInterfaceError::RateLimiterUpdateFailed(
-                devices::Error::FailedReadTap
-            )),
-            ErrorKind::Internal
-        );
-        assert_eq!(
-            error_kind(NetworkInterfaceError::UpdateNotAllowedPostBoot),
-            ErrorKind::User
-        );
-    }
-
-    #[test]
-    #[allow(clippy::cognitive_complexity)]
-    fn test_start_microvm_error_conversion_cl() {
-        // Test `StartMicrovmError` conversion.
-        assert_eq!(
-            error_kind(StartMicrovmError::ConfigureVm(
-                vstate::Error::NotEnoughMemorySlots
-            )),
-            ErrorKind::Internal
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::CreateBlockDevice(
-                io::Error::from_raw_os_error(0)
-            )),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::CreateNetDevice(
-                devices::virtio::Error::TapOpen(utils::net::TapError::CreateTap(
-                    io::Error::from_raw_os_error(0)
-                ))
-            )),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::CreateRateLimiter(
-                io::Error::from_raw_os_error(0)
-            )),
-            ErrorKind::Internal
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::CreateVsockBackend(
-                devices::virtio::vsock::VsockUnixBackendError::InvalidPortRequest
-            )),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::CreateVsockDevice(
-                devices::virtio::vsock::VsockError::NoData
-            )),
-            ErrorKind::Internal
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::GuestMemory(
-                memory_model::GuestMemoryError::NoMemoryRegions
-            )),
-            ErrorKind::Internal
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::KernelCmdline(String::new())),
-            ErrorKind::User
-        );
-        #[cfg(target_arch = "x86_64")]
-        assert_eq!(
-            error_kind(StartMicrovmError::LegacyIOBus(
-                device_manager::legacy::Error::EventFd(io::Error::from_raw_os_error(0))
-            )),
-            ErrorKind::Internal
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::LoadCommandline(
-                kernel::cmdline::Error::CommandLineOverflow
-            )),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::LoadCommandline(
-                kernel::cmdline::Error::CommandLineCopy
-            )),
-            ErrorKind::Internal
-        );
-    }
-
-    #[test]
-    fn test_start_microvm_error_conversion_mv() {
-        assert_eq!(
-            error_kind(StartMicrovmError::MicroVMAlreadyRunning),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::MissingKernelConfig),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::NetDeviceNotConfigured),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::OpenBlockDevice(
-                io::Error::from_raw_os_error(0)
-            )),
-            ErrorKind::User
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::RegisterBlockDevice(
-                device_manager::mmio::Error::IrqsExhausted
-            )),
-            ErrorKind::Internal
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::RegisterNetDevice(
-                device_manager::mmio::Error::IrqsExhausted
-            )),
-            ErrorKind::Internal
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::RegisterMMIODevice(
-                device_manager::mmio::Error::IrqsExhausted
-            )),
-            ErrorKind::Internal
-        );
-        assert_eq!(
-            error_kind(StartMicrovmError::RegisterVsockDevice(
-                device_manager::mmio::Error::IrqsExhausted
-            )),
-            ErrorKind::Internal
-        );
-    }
-
-    #[test]
-    #[allow(clippy::cognitive_complexity)]
-    fn test_error_messages() {
-        // Enum `Error`
-
-        #[cfg(target_arch = "x86_64")]
-        assert_eq!(
-            format!(
-                "{:?}",
-                Error::CreateLegacyDevice(device_manager::legacy::Error::EventFd(
-                    io::Error::from_raw_os_error(42)
-                ))
-            ),
-            format!(
-                "Error creating legacy device: EventFd({:?})",
-                io::Error::from_raw_os_error(42)
-            )
-        );
-        assert_eq!(
-            format!("{:?}", Error::EpollFd(io::Error::from_raw_os_error(42))),
-            "Epoll fd error: No message of desired type (os error 42)"
-        );
-        assert_eq!(
-            format!("{:?}", Error::EventFd(io::Error::from_raw_os_error(42))),
-            "Event fd error: No message of desired type (os error 42)"
-        );
-        assert_eq!(
-            format!("{:?}", Error::DeviceEventHandlerNotFound),
-            "Device event handler not found. This might point to a guest device driver issue."
-        );
-        assert_eq!(
-            format!("{:?}", Error::DeviceEventHandlerInvalidDowncast),
-            "Device event handler couldn't be downcasted to expected type."
-        );
-        assert_eq!(
-            format!("{:?}", Error::KvmContext(vstate::Error::KvmApiVersion(1))),
-            "Failed to validate KVM support: KvmApiVersion(1)"
-        );
-        assert_eq!(
-            format!("{:?}", Error::Poll(io::Error::from_raw_os_error(42))),
-            "Epoll wait failed: No message of desired type (os error 42)"
-        );
-        assert_eq!(
-            format!("{:?}", Error::Serial(io::Error::from_raw_os_error(42))),
-            format!(
-                "Error writing to the serial console: {:?}",
-                io::Error::from_raw_os_error(42)
-            )
-        );
-        assert_eq!(
-            format!("{:?}", Error::TimerFd(io::Error::from_raw_os_error(42))),
-            "Error creating timer fd: No message of desired type (os error 42)"
-        );
-        assert_eq!(
-            format!("{:?}", Error::Vm(vstate::Error::HTNotInitialized)),
-            "Error opening VM fd: HTNotInitialized"
-        );
-
-        // Enum `ErrorKind`
-
-        assert_ne!(ErrorKind::User, ErrorKind::Internal);
-        assert_eq!(format!("{:?}", ErrorKind::User), "User");
-        assert_eq!(format!("{:?}", ErrorKind::Internal), "Internal");
-
-        // Enum VmmActionError
-
-        assert_eq!(
-            format!(
-                "{:?}",
-                VmmActionError::BootSource(
-                    ErrorKind::User,
-                    BootSourceConfigError::InvalidKernelCommandLine(
-                        kernel::cmdline::Error::HasSpace.to_string()
-                    )
-                )
-            ),
-            "BootSource(User, InvalidKernelCommandLine(\"Command line string contains a space\"))"
-        );
-        assert_eq!(
-            format!(
-                "{:?}",
-                VmmActionError::DriveConfig(
-                    ErrorKind::User,
-                    DriveError::BlockDevicePathAlreadyExists
-                )
-            ),
-            "DriveConfig(User, BlockDevicePathAlreadyExists)"
-        );
-        assert_eq!(
-            format!(
-                "{}",
-                VmmActionError::Logger(
-                    ErrorKind::User,
-                    LoggerConfigError::FlushMetrics(String::from("Failed to flush metrics"))
-                )
-            ),
-            "Failed to flush metrics"
-        );
-        assert_eq!(
-            format!(
-                "{:?}",
-                VmmActionError::Logger(
-                    ErrorKind::User,
-                    LoggerConfigError::FlushMetrics(String::from("foobar"))
-                )
-            ),
-            "Logger(User, FlushMetrics(\"foobar\"))"
-        );
-        assert_eq!(
-            format!(
-                "{}",
-                VmmActionError::Logger(
-                    ErrorKind::User,
-                    LoggerConfigError::InitializationFailure(String::from(
-                        "Failed to initialize logger"
-                    ))
-                )
-            ),
-            "Failed to initialize logger"
-        );
-        assert_eq!(
-            format!(
-                "{:?}",
-                VmmActionError::Logger(
-                    ErrorKind::User,
-                    LoggerConfigError::InitializationFailure(String::from("foobar"))
-                )
-            ),
-            "Logger(User, InitializationFailure(\"foobar\"))"
-        );
-        assert_eq!(
-            format!(
-                "{:?}",
-                VmmActionError::MachineConfig(ErrorKind::User, VmConfigError::InvalidMemorySize)
-            ),
-            "MachineConfig(User, InvalidMemorySize)"
-        );
-        assert_eq!(
-            format!(
-                "{:?}",
-                VmmActionError::NetworkConfig(
-                    ErrorKind::User,
-                    NetworkInterfaceError::DeviceIdNotFound
-                )
-            ),
-            "NetworkConfig(User, DeviceIdNotFound)"
-        );
-        assert_eq!(
-            format!(
-                "{:?}",
-                VmmActionError::SendCtrlAltDel(
-                    ErrorKind::User,
-                    I8042Error(I8042DeviceError::InternalBufferFull)
-                )
-            ),
-            "SendCtrlAltDel(User, InternalBufferFull)"
-        );
-        assert_eq!(
-            format!(
-                "{}",
-                VmmActionError::SendCtrlAltDel(
-                    ErrorKind::User,
-                    I8042Error(I8042DeviceError::InternalBufferFull)
-                )
-            ),
-            I8042DeviceError::InternalBufferFull.to_string()
-        );
-        assert_eq!(
-            VmmActionError::SendCtrlAltDel(
-                ErrorKind::User,
-                I8042Error(I8042DeviceError::InternalBufferFull)
-            )
-            .kind(),
-            &ErrorKind::User
-        );
-        assert_eq!(
-            format!(
-                "{:?}",
-                VmmActionError::VsockConfig(ErrorKind::User, VsockError::UpdateNotAllowedPostBoot)
-            ),
-            "VsockConfig(User, UpdateNotAllowedPostBoot)"
-        );
-        assert_eq!(
-            format!(
-                "{}",
-                VmmActionError::VsockConfig(ErrorKind::User, VsockError::UpdateNotAllowedPostBoot)
-            ),
-            "The update operation is not allowed after boot."
-        );
-    }
-}
-*/
