+++ src/vmm/src/lib.rs
@@ -40,7 +40,6 @@ pub mod controller;
 /// Syscalls allowed through the seccomp filter.
 pub mod default_syscalls;
 pub(crate) mod device_manager;
-pub mod error;
 /// Resource store for configured microVM resources.
 pub mod resources;
 /// microVM RPC API adapters.
@@ -52,6 +51,7 @@ pub mod vmm_config;
 mod vstate;
 
 use std::collections::HashMap;
+use std::fmt::{Display, Formatter};
 use std::io;
 use std::os::unix::io::{AsRawFd, RawFd};
 use std::sync::mpsc::{channel, Receiver, Sender};
@@ -68,13 +68,13 @@ use device_manager::mmio::MMIODeviceInfo;
 use device_manager::mmio::MMIODeviceManager;
 use devices::virtio::EpollConfigConstructor;
 use devices::{BusDevice, DeviceEventT, EpollHandler, RawIOHandler};
-use error::{Error, Result};
 use kernel::cmdline::Cmdline as KernelCmdline;
+use logger::error::LoggerError;
 #[cfg(target_arch = "x86_64")]
 use logger::LogOption;
 use logger::{Metric, LOGGER, METRICS};
 use memory_model::{GuestAddress, GuestMemory};
-use polly::event_manager::EventManager;
+use polly::event_manager::{self, EventManager};
 use utils::eventfd::EventFd;
 use utils::terminal::Terminal;
 use utils::time::TimestampUs;
@@ -356,6 +356,103 @@ impl Drop for EpollContext {
     }
 }
 
+/// Errors associated with the VMM internal logic. These errors cannot be generated by direct user
+/// input, but can result from bad configuration of the host (for example if Firecracker doesn't
+/// have permissions to open the KVM fd).
+#[derive(Debug)]
+pub enum Error {
+    /// This error is thrown by the minimal boot loader implementation.
+    ConfigureSystem(arch::Error),
+    /// Legacy devices work with Event file descriptors and the creation can fail because
+    /// of resource exhaustion.
+    #[cfg(target_arch = "x86_64")]
+    CreateLegacyDevice(device_manager::legacy::Error),
+    /// An operation on the epoll instance failed due to resource exhaustion or bad configuration.
+    EpollFd(io::Error),
+    /// Cannot read from an Event file descriptor.
+    EventFd(io::Error),
+    /// An event arrived for a device, but the dispatcher can't find the event (epoll) handler.
+    DeviceEventHandlerNotFound,
+    /// An epoll handler can't be downcasted to the desired type.
+    DeviceEventHandlerInvalidDowncast,
+    /// Cannot access kernel file.
+    KernelFile(io::Error),
+    /// Cannot open /dev/kvm. Either the host does not have KVM or Firecracker does not have
+    /// permission to open the file descriptor.
+    KvmContext(vstate::Error),
+    #[cfg(target_arch = "x86_64")]
+    /// Cannot add devices to the Legacy I/O Bus.
+    LegacyIOBus(device_manager::legacy::Error),
+    /// Cannot load command line.
+    LoadCommandline(kernel::cmdline::Error),
+    /// Internal logger error.
+    Logger(LoggerError),
+    /// I8042 Error.
+    I8042Error(devices::legacy::I8042DeviceError),
+    /// Epoll wait failed.
+    Poll(io::Error),
+    /// Cannot add a device to the MMIO Bus.
+    RegisterMMIODevice(device_manager::mmio::Error),
+    /// Cannot build seccomp filters.
+    SeccompFilters(seccomp::Error),
+    /// Write to the serial console failed.
+    Serial(io::Error),
+    /// Cannot set mode for terminal.
+    StdinHandle(utils::errno::Error),
+    /// Cannot create Timer file descriptor.
+    TimerFd(io::Error),
+    /// Vcpu error.
+    Vcpu(vstate::Error),
+    /// Cannot spawn a new Vcpu thread.
+    VcpuSpawn(std::io::Error),
+    /// Vm error.
+    Vm(vstate::Error),
+    /// Polly error wrapper.
+    EventManager(event_manager::Error),
+}
+
+impl Display for Error {
+    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
+        use self::Error::*;
+
+        match self {
+            ConfigureSystem(e) => write!(f, "System configuration error: {:?}", e),
+            #[cfg(target_arch = "x86_64")]
+            CreateLegacyDevice(e) => write!(f, "Error creating legacy device: {:?}", e),
+            EpollFd(e) => write!(f, "Epoll fd error: {}", e),
+            EventFd(e) => write!(f, "Event fd error: {}", e),
+            DeviceEventHandlerNotFound => write!(
+                f,
+                "Device event handler not found. This might point to a guest device driver issue."
+            ),
+            DeviceEventHandlerInvalidDowncast => write!(
+                f,
+                "Device event handler couldn't be downcasted to expected type."
+            ),
+            KernelFile(e) => write!(f, "Cannot access kernel file: {}", e),
+            KvmContext(e) => write!(f, "Failed to validate KVM support: {:?}", e),
+            #[cfg(target_arch = "x86_64")]
+            LegacyIOBus(e) => write!(f, "Cannot add devices to the legacy I/O Bus. {}", e),
+            LoadCommandline(e) => write!(f, "Cannot load command line: {}", e),
+            Logger(e) => write!(f, "Logger error: {}", e),
+            I8042Error(e) => write!(f, "I8042 error: {}", e),
+            Poll(e) => write!(f, "Epoll wait failed: {}", e),
+            RegisterMMIODevice(e) => write!(f, "Cannot add a device to the MMIO Bus. {}", e),
+            SeccompFilters(e) => write!(f, "Cannot build seccomp filters: {}", e),
+            Serial(e) => write!(f, "Error writing to the serial console: {:?}", e),
+            StdinHandle(e) => write!(f, "Failed to set mode for terminal: {}", e),
+            TimerFd(e) => write!(f, "Error creating timer fd: {}", e),
+            Vcpu(e) => write!(f, "Vcpu error: {}", e),
+            VcpuSpawn(e) => write!(f, "Cannot spawn Vcpu thread: {}", e),
+            Vm(e) => write!(f, "Vm error: {}", e),
+            EventManager(e) => write!(f, "Event manager error: {:?}", e),
+        }
+    }
+}
+
+/// Shorthand result type for internal VMM commands.
+pub type Result<T> = std::result::Result<T, Error>;
+
 /// Encapsulates configuration parameters for the guest vCPUS.
 pub struct VcpuConfig {
     /// Number of guest VCPUs.
@@ -442,7 +539,9 @@ impl Vmm {
 
     #[cfg(target_arch = "aarch64")]
     fn setup_interrupt_controller(&mut self) -> Result<()> {
-        self.vm.setup_irqchip(self.vcpu_config.vcpu_count)
+        self.vm
+            .setup_irqchip(self.vcpu_config.vcpu_count)
+            .map_err(Error::Vm)
     }
 
     #[cfg(target_arch = "x86_64")]
